+++ scripts/import.sh
@@ -56,20 +56,15 @@ load_files() {
       local target="${file/#~/"$HOME"}"
       unset local cmd_prefix
 
-      if [[ ! -d "$source" && ! -f "$source" ]]; then
-        echo "[ WARN ] Invalid file to import: $file [ source folder: $data_folder ]" && continue
-      fi
+      [[ ! -d "$source" && ! -f "$source" ]] \
+          && echo "[ WARN ] Invalid file to import: $file [ source folder: $data_folder ]" && continue
 
       local target_parent_dir="$(dirname "$target")"
 
       dir_permission_check "$target_parent_dir" || local cmd_prefix="sudo"
 
-      if [[ -d "$source" ]]; then
-        $cmd_prefix mkdir -p "$target" && $cmd_prefix rsync -a "$source"/* "$target"
-      fi
-      if [[ -f "$source" ]]; then
-        $cmd_prefix mkdir -p "$target_parent_dir" && $cmd_prefix cp "$source" "$target"
-      fi
+      [[ -d "$source" ]] && $cmd_prefix mkdir -p "$target" && $cmd_prefix rsync -a "$source"/* "$target"
+      [[ -f "$source" ]] && $cmd_prefix mkdir -p "$target_parent_dir" && $cmd_prefix cp "$source" "$target"
     done
 
   done < <(jq -cr "$jq_filter" "$config_json")
